build: return to js
