+++ backend/.env
@@ -1,2 +1,3 @@
 API_KEY='ADC7D7B1D29F4DF3974FCA5656071938'
 BASE_URL='https://occupational-maps-api.instituteforapprenticeships.org/api/v1'
+PORT=8080
