+++ backend/.gitignore
@@ -8,6 +8,7 @@
 
 # misc
 .DS_Store
+.env
 .env.local
 .env.development.local
 .env.test.local
