+++ backend/package.json
@@ -1,23 +1,21 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "proxy server for pathways backend ",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "alphonso",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/node": "^20.14.9",
-    "@types/node-fetch": "^2.6.11",
-    "dotenv": "^16.4.5"
-  },
-  "dependencies": {
-    "express": "^4.19.2",
-    "node-fetch": "^3.3.2",
-    "prettier": "^3.3.2"
-  }
+    "name": "backend",
+    "version": "1.0.0",
+    "description": "proxy server for pathways backend ",
+    "main": "index.ts",
+    "type": "module",
+    "scripts": {
+        "dev": "node src/index.js"
+    },
+    "author": "alphonso",
+    "license": "ISC",
+    "devDependencies": {
+        "cors": "^2.8.5",
+        "dotenv": "^16.4.5"
+    },
+    "dependencies": {
+        "express": "^4.19.2",
+        "node-fetch": "^3.3.2",
+        "prettier": "^3.3.2"
+    }
 }
