+++ backend/src/index.js
@@ -1,0 +1,5 @@
+import server from "../server/server.js"
+
+server.listen(process.env.PORT, () =>
+    console.log("Server is live!")
+)
